made user folder and moved user components into it
