+++ 3sp/src/app/User/types/courseTypes.ts
